refractor: homologación de respuestas de los endpoints de locaciones.
- Se actualizaron los métodos del servicio y controlador asociado a las
  locaciones para que las respuestas de la api se envién según la
  estructura de la clase "ResponseServices".
- Se realizaron algunos pequeños ajustes al modelo y al DTO de
  locaciones para que siga la misma estructura utiizada en otros modelos
  del proyecto.
- Se actualizaron los casos de pruebas para que tomen en cuenta la nueva
  estructura de respuestas del servicio de locaciones.

+++ src/main/java/com/futuro/api_iot_data/repositories/LocationRepository.java
@@ -9,7 +9,7 @@ import com.futuro.api_iot_data.models.Location;
  * Proporciona métodos para acceder y gestionar datos relacionados con locaciones en la base de datos.
  */
 @Repository
-public interface LocationRepository extends JpaRepository<Location,Long>{
+public interface LocationRepository extends JpaRepository<Location,Integer>{
 	/**
      * Verifica si existe una locación con el nombre especificado.
      *
@@ -28,5 +28,5 @@ public interface LocationRepository extends JpaRepository<Location,Long>{
      * @return {@code true} si existe una locación con el nombre proporcionado (excluyendo la locación con el ID dado),
      *         {@code false} en caso contrario.
      */
-	boolean existsByLocationNameAndLocationIdNot(String LocationName, Long locationId);
+	boolean existsByLocationNameAndLocationIdNot(String LocationName, Integer locationId);
 }
